Update to select mesh + set item tag.
First stage of setting the item tag and selecting the mesh.

Having the mesh selected makes the pbr loader set the Txt locator to UV map.  (a mesh needs to be available to choose the map)

With the item selected, we can then select the material mask and set the item tag.

This splits the process from one bug old quee item, into several smaller ones.
First - import any meshes
Then Select meshes
Then Import Materials
Then select and set masks.

This makes it slightly slower, as it has to wait for each stage to finsih to do the next one, which is currently just on a timer.