Change to the bridge to make it Set the UV and the Mask for imports
Changed how the bridge works to treat geometry and materials as separate tasks.
If importing geometry, you can now choose to select the imported mesh. If a mesh is selected, when imported, the in built pbr loader will set the UV map to currently chosen UV map.
There is also a flag to set the item mask on the mesh. This will take the first selected mesh, and set the mask on the new mateiral to that mesh. (note that if the import brings in multiple meshes, this means only one of those meshes will have the material applied.)

The bridge now basically makes a mini-queue for each push from quixel:
* geometry import
* select meshes
* import mateirals
* set material masks

The two intervals are for controlling this. The Main interval is how long to check for input from quixel. The stage interval is how long to wait between stages. Once all stages are complete, it goes back to waiting for the main interval to check again. I have exposed these values.